+++ CHANGELOG.md
@@ -5,7 +5,7 @@
 ### Code changes:
 - Fixed a problem with the animation timer for the loading Forge objects menu
 - Added a new crosshair state that prevents you to place objects in a prohibited area
-- Fixed a problem with the count of buttons displayed in the map voti ng menu
+- Fixed a problem with the count of buttons displayed in the map voting menu
 - Added a debug message to print how many objects exist on the Forge objects database
 - Added a new format to save the name of all the Forge maps using a standard convention with underscores
 - Fixed a bug with a debug feature that allowed to spawn test bipeds